@@ -1,6 +1,6 @@
 {
   "name": "profile-demo",
-  "version": "1.3.5",
+  "version": "1.3.6",
   "main": "index.js",
   "author": "ArcBlock",
   "license": "MIT",
@@ -19,13 +19,16 @@
     "coverage": "npm run test -- --coverage"
   },
   "dependencies": {
-    "@arcblock/did-connect": "^1.17.18",
-    "@arcblock/ux": "^1.17.18",
-    "@blocklet/sdk": "1.7.18",
-    "@blocklet/ui-react": "^1.17.18",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.57",
+    "@arcblock/did-connect": "^2.1.7",
+    "@arcblock/ux": "^2.1.7",
+    "@blocklet/sdk": "1.7.19",
+    "@blocklet/ui-react": "^2.1.7",
+    "@mui/material": "^5.8.1",
+    "@mui/styles": "^5.8.0",
+    "@emotion/styled": "^11.8.1",
+    "@emotion/react": "^11.9.0",
+    "@mui/icons-material": "^5.8.0",
+    "@mui/lab": "^5.0.0-alpha.83",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
@@ -42,10 +45,13 @@
     "fs-extra": "^9.1.0",
     "lodash": "^4.17.20",
     "morgan": "^1.10.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-scripts": "4.0.2",
-    "styled-components": "^5.2.1"
+    "styled-components": "^5.3.5"
+  },
+  "resolutions": {
+    "styled-components": "5.3.5"
   },
   "eslintConfig": {
     "extends": [
