@@ -1,7 +1,7 @@
 import React from 'react';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import { MuiThemeProvider } from '@material-ui/core/styles';
-import { ThemeProvider } from 'styled-components';
+import CssBaseline from '@mui/material/CssBaseline';
+import { ThemeProvider, StyledEngineProvider } from '@mui/material/styles';
+import { ThemeProvider as StyledThemeProvider } from 'styled-components';
 import { LocaleProvider } from '@arcblock/ux/lib/Locale/context';
 import { create } from '@arcblock/ux/lib/Theme';
 
@@ -37,16 +37,18 @@ function App() {
   });
 
   return (
-    <MuiThemeProvider theme={theme}>
+    <StyledEngineProvider injectFirst>
       <ThemeProvider theme={theme}>
-        <LocaleProvider translations={translations}>
-          <SessionProvider serviceHost={prefix}>
-            <CssBaseline />
-            <Main />
-          </SessionProvider>
-        </LocaleProvider>
+        <StyledThemeProvider theme={theme}>
+          <LocaleProvider translations={translations}>
+            <SessionProvider serviceHost={prefix}>
+              <CssBaseline />
+              <Main />
+            </SessionProvider>
+          </LocaleProvider>
+        </StyledThemeProvider>
       </ThemeProvider>
-    </MuiThemeProvider>
+    </StyledEngineProvider>
   );
 }
 
